+++ classes/Property.php
@@ -1,10 +1,10 @@
 <?php
 
+/** @noinspection PhpDocMissingThrowsInspection */
+/** @noinspection PhpUnhandledExceptionInspection */
+
 namespace Neat\Object;
 
-use DateTime;
-use DateTimeImmutable;
-use DateTimeInterface;
 use ReflectionProperty;
 
 class Property
@@ -23,20 +23,15 @@ class Property
      * Property constructor
      *
      * @param ReflectionProperty $reflection
+     * @param string             $type
      * @note Activates the reflection's accessible flag
      */
-    public function __construct(ReflectionProperty $reflection)
+    public function __construct(ReflectionProperty $reflection, string $type = null)
     {
         $reflection->setAccessible(true);
 
         $this->reflection = $reflection;
-
-        if (preg_match('/\\s@var\\s([\\w\\\\]+)(?:\\|null)?\\s/', $reflection->getDocComment(), $matches)) {
-            $this->type = strtr(ltrim($matches[1], '\\'), [
-                'integer' => 'int',
-                'boolean' => 'bool',
-            ]);
-        }
+        $this->type       = $type;
     }
 
     /**
@@ -44,9 +39,29 @@ class Property
      *
      * @return string
      */
-    public function name()
+    public function name(): string
     {
         return $this->reflection->getName();
+    }
+
+    /**
+     * Is static?
+     *
+     * @return bool
+     */
+    public function static(): bool
+    {
+        return $this->reflection->isStatic();
+    }
+
+    /**
+     * Get doc comment
+     *
+     * @return string
+     */
+    public function comment(): string
+    {
+        return $this->reflection->getDocComment() ?: '';
     }
 
     /**
@@ -57,6 +72,28 @@ class Property
     public function type()
     {
         return $this->type;
+    }
+
+    /**
+     * Cast value to string
+     *
+     * @param mixed $value
+     * @return string
+     */
+    public function toString($value): string
+    {
+        return (string) $value;
+    }
+
+    /**
+     * Cast value from string
+     *
+     * @param string $value
+     * @return mixed
+     */
+    public function fromString(string $value)
+    {
+        return $value;
     }
 
     /**
@@ -72,21 +109,7 @@ class Property
             return null;
         }
 
-        switch ($this->type) {
-            case 'bool':
-                return $value ? 1 : 0;
-            case 'int':
-                return (int)$value;
-            case 'DateTime':
-            case 'DateTimeImmutable':
-                if (!$value instanceof DateTimeInterface) {
-                    $value = new DateTime($value);
-                }
-
-                return $value->format('Y-m-d H:i:s');
-            default:
-                return $value;
-        }
+        return $this->toString($value);
     }
 
     /**
@@ -98,40 +121,9 @@ class Property
     public function set($object, $value)
     {
         if ($value !== null) {
-            switch ($this->type) {
-                case 'bool':
-                case 'int':
-                    settype($value, $this->type);
-                    break;
-                case 'DateTime':
-                    $value = new DateTime($value);
-                    break;
-                case 'DateTimeImmutable':
-                    $value = new DateTimeImmutable($value);
-                    break;
-            }
+            $value = $this->fromString($value);
         }
 
         $this->reflection->setValue($object, $value);
-    }
-
-    /**
-     * Is static?
-     *
-     * @return bool
-     */
-    public function static()
-    {
-        return $this->reflection->isStatic();
-    }
-
-    /**
-     * Get doc block
-     *
-     * @return bool|string
-     */
-    public function docBlock()
-    {
-        return $this->reflection->getDocComment();
     }
 }