+++ changelog.md
@@ -1,10 +1,12 @@
 # Changelog
-All notable changes to Neat Database components will be documented in this file.
+All notable changes to Neat Object components will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Removed property type method.
 
 ## [0.9.3] - 2019-10-01
 ### Fixed